Add evading timer
After having no predator detected, preys have a cool down time before changing to wander state. In this time, prey will continue to move with the direction against the last predator it saw. This cool down timer is reseted if prey encounter a new predator. This help to reduce jittering behavior of prey.