Fix animals Prefabs code for instantiation during runtime
Before, when instantiating new Animal with Prefabs during runtime, the start
script was ran instantly before the Prefabs code finishes running and
that was a problem because Animal.start expect children game object
before the Prefabs code to create the children was ran.

Move the code to set the parent for the new prefab down below its
children creation to avoid this problem. (start code is ran when setting
the parent if the parent is started).